This may be a better approach
In this code I use the /** string as a separator
between blocks of code. After splitting the string
with this separator I throw out any chunks that
don't contain a ';', because that seems to be a
character that is present in every valid block.

Lastly, I output the blocks one-by-one. Since split
removes the separator string /** during the process
of splitting the original list_of_things, I add it
back to each block using the string + operator.